Adding apache vhost config templates: SVN, CF.
SVN is based on my other dev VM and just shares any SVN repo in the directory.
I'm not sure if this will work properly with repos named project/repo or
user/repo.

CF is based on the default rails template, which has been partially updated for
HTTPS, but needs a lot of work.