Forking the Vagrantfile.
Vagrantfile-cf is for the codefoundry application and Vagrantfile-ruby-1.9.2 is
for building a ruby-1.9.2 based box.  This is handy because it takes a *long*
time to compile and get everything set up and it's nice to not have to do that
every time you want to destroy the current VM.  The real reason, however, is
that the provisioning process needs to happen twice, so that chef is run a
second time under 1.9.2.  This makes all of the gem_package calls install gems
to 1.9.2, including bundler.  Without this change, everything is installed to
system ruby and it all breaks when trying to bring up apache/passenger.