Updating to the latest in my working copy.
Should have been checking in more snaps, but I haven't worked on it in a while.
I am at the point where it will bring up a full rails stack and run the
application, but I need to add the db:migrate step.  There may be more bugs
after that.